microservices/student-profiler/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pandas==2.2.3
scikit-learn==1.3.2
sqlalchemy==2.0.19
psycopg2-binary==2.9.7
pyyaml==6.0
joblib==1.4.1
pytest==7.4.2
httpx==0.24.1
python-dotenv==1.0.0

